Change theory function name field to string field.
Clingo allows theory function names to be symbols like + and - (in
fact, that is how parsed theory terms are represented, e.g. the
unparsed theory term "a + b" is parsed into a theory function with
name + and arguments a, b. Therefore, restricting the name field to be
a constant field (as it was before this change) would be incorrect.